infra: Add steps to get and include Qermit docs in prototype
Right now I've just copy-pasted the steps we did for Lambeq and changed
them to refer to isobelhooper/Qermit and files with "qermit" in instead.

I don't think it's feasible to maintain these as copy-pasted for everything,
but I'll need to see what I can do with matrix: as that might not work
exactly how we'd like.