ci: Use a local action to share docs extraction code
Tried this with matrix strategy to start with, but that requires
separate jobs for the matrix part and the "deploy it all to Github Pages"
part, and workspaces are only shared within a job, not within a workflow.

So we'd have to cache the extracted tarballs between jobs - which requires
zipping them up again - and that seemed a bit much.

Instead, I've extracted the shared code into an action in this repo,
and we have an entry for each of the included products where we extract
their docs into the correct subdirectory.