Remove Qermit and Lambeq from the unified docs build
This is at Ross's request, as they aren't the products that we are
prioritising the most for the unified docs.

Also, the versions that were used were from forks of those repos, and so
were not up to date. Removing them means that people don't find these
versions (which weren't up to date) if they're looking for the actual docs
for Qermit or Lambeq.