Swap around 'Get Started' and 'Tutorials and Documentation' links
It looks like the first link in each card was to the overall docs site, and the second link was to a section about how to get started - but the link text implied it was the opposite way around. This commit swaps the link targets around so that the link target and text match.